+++ src/index.ts
@@ -597,19 +597,7 @@ export class Scru128Generator {
    *   was broken because the latest timestamp was less than the previous one by
    *   ten seconds or more.
    *
-   * @example
-   * ```javascript
-   * import { Scru128Generator } from "scru128";
-   *
-   * const g = new Scru128Generator();
-   * const x = g.generate();
-   * const y = g.generate();
-   * if (g.getLastStatus() === "CLOCK_ROLLBACK") {
-   *   throw new Error("clock moved backward");
-   * } else {
-   *   console.assert(x.compareTo(y) < 0);
-   * }
-   * ```
+   * @deprecated Use {@link generateOrAbort} to guarantee monotonicity.
    */
   getLastStatus() {
     return this.lastStatus;